chore: add dialog component
Dialog component was added to aid user interaction.
