+++ src/components/Dialog/index.tsx
@@ -1,0 +1,1 @@
+export { default as Dialog } from "./Dialog";
